Fix blocking and make general improvements
Non-asynchronous functions means that the entire bot stops when the
function is running, meaning the bot is blocked from doing anything else
while, for example, calculating inventory. This makes the bot less
responsive especially if it's being used in multiple places at once.
This is making them all asynchronous, meaning the bot will no longer be
blocked while running certain functions.

General code improvements include making functions for what used to be
if statements. The most notable change to the user is that item levels
are now more consistent, so instead of having the Stone level be "Axe"
it's now less-confusingly "Stone Axe". This is also how the item level
detection now works, so not having the material in front of the item
breaks the get_item_level function.